+++ src/torchCompactRadius/cppSrc/neighborhoodDynamic.cpp
@@ -10,7 +10,7 @@
  * @param t The input tensor.
  * @param name The name of the accessor.
  * @param cuda Flag indicating whether the tensor should be on CUDA.
- * @param verbose Flag indicating whether to print verbose information.
+ * @param verbose Flag indicating whether to print64_t verbose information.
  * @param optional Flag indicating whether the tensor is optional.
  * @return The packed accessor for the tensor.
  * @throws std::runtime_error If the tensor is not defined (and not optional), not contiguous, not on CUDA (if cuda=true), or has an incorrect dimension.
@@ -60,14 +60,14 @@ auto getAccessor(const torch::Tensor &t, const std::string &name, bool cuda = fa
  * @param minDomain_ The minimum domain bounds.
  * @param periodicity_ The periodicity flags.
  * @param mode The support mode.
- * @param verbose Flag indicating whether to print verbose information.
+ * @param verbose Flag indicating whether to print64_t verbose information.
  * @return The number of neighbors for each particle in the query set.
  */
 template<typename float_t = float>
 torch::Tensor countNeighbors_t(
-    torch::Tensor queryPositions_, torch::Tensor querySupport_, int searchRange, 
+    torch::Tensor queryPositions_, torch::Tensor querySupport_, int32_t searchRange, 
     torch::Tensor sortedPositions_, torch::Tensor sortedSupport_,
-    torch::Tensor hashTable_, int hashMapLength, 
+    torch::Tensor hashTable_, int32_t hashMapLength, 
     torch::Tensor numCells_, torch::Tensor cellTable_,
     torch::Tensor qMin_, float_t hCell, torch::Tensor maxDomain_, torch::Tensor minDomain_, torch::Tensor periodicity_,
     std::string mode, bool verbose = false){
@@ -94,13 +94,13 @@ torch::Tensor countNeighbors_t(
     auto sortedSupport = getAccessor<float_t, 1>(sortedSupport_, "sortedSupport", useCuda, verbose, supportMode::gather == searchMode);
 
     // Get the dimensions of the input tensors
-    int nQuery = queryPositions.size(0);
-    int dim = queryPositions.size(1);
-    int nSorted = sortedPositions.size(0);
+    int32_t nQuery = queryPositions.size(0);
+    int32_t dim = queryPositions.size(1);
+    int32_t nSorted = sortedPositions.size(0);
     
     // Check if the datastructure tensors are defined and contiguous and have the correct dimensions
-    auto hashTable = getAccessor<int, 2>(hashTable_, "hashTable", useCuda, verbose);
-    auto numCells = getAccessor<int, 1>(numCells_, "numCells", useCuda, verbose);
+    auto hashTable = getAccessor<int64_t, 2>(hashTable_, "hashTable", useCuda, verbose);
+    auto numCells = getAccessor<int32_t, 1>(numCells_, "numCells", useCuda, verbose);
     auto cellTable = getAccessor<int64_t, 2>(cellTable_, "cellTable", useCuda, verbose);
     auto qMin = getAccessor<float_t, 1>(qMin_, "qMin", useCuda, verbose);
     auto maxDomain = getAccessor<float_t, 1>(maxDomain_, "maxDomain", useCuda, verbose);
@@ -176,18 +176,18 @@ torch::Tensor countNeighbors_t(
         }
     }
     // Allocate output tensor for the neighbor counters
-    auto neighborCounters = torch::zeros({nQuery}, defaultOptions.dtype(torch::kInt32));
+    auto neighborCounters = torch::zeros({nQuery}, defaultOptions.dtype(torch::kInt64));
 
     // Create the accessors for the input tensors as packed accessors
     auto queryPositionAccessor = queryPositions_.packed_accessor32<float_t, 2, traits>();
     auto querySupportAccessor = querySupport_.packed_accessor32<float_t, 1, traits>();
     auto referencePositionAccessor = sortedPositions_.packed_accessor32<float_t, 2, traits>();
     auto referenceSupportAccessor = sortedSupport_.packed_accessor32<float_t, 1, traits>();
-    auto hashTableAccessor = hashTable_.packed_accessor32<int32_t, 2, traits>();
+    auto hashTableAccessor = hashTable_.packed_accessor32<int64_t, 2, traits>();
     auto celTableAccessor = cellTable_.packed_accessor32<int64_t, 2, traits>();
     auto offsetAccessor = offsets.packed_accessor32<int32_t, 2, traits>();
     auto numCellsAccessor = numCells_.packed_accessor32<int32_t, 1, traits>();
-    auto neighborCounterAccessor = neighborCounters.packed_accessor32<int32_t, 1, traits>();
+    auto neighborCounterAccessor = neighborCounters.packed_accessor32<int64_t, 1, traits>();
 
     // Loop over all query particles and count the number of neighbors per particle
     if(queryPositions_.is_cuda()){
@@ -252,15 +252,15 @@ torch::Tensor countNeighbors_t(
  * @param minDomain_ The minimum domain bounds.
  * @param periodicity_ The periodicity flags.
  * @param mode The support mode.
- * @param verbose Flag indicating whether to print verbose information.
+ * @param verbose Flag indicating whether to print64_t verbose information.
  * @return The neighbor list as a pair of tensors
  */
 template<typename float_t = float>
 std::pair<torch::Tensor, torch::Tensor> buildNeighborList_t(
-    torch::Tensor neighborCounter_, torch::Tensor neighborOffsets_, int neighborListLength,
-    torch::Tensor queryPositions_, torch::Tensor querySupport_, int searchRange, 
+    torch::Tensor neighborCounter_, torch::Tensor neighborOffsets_, int32_t neighborListLength,
+    torch::Tensor queryPositions_, torch::Tensor querySupport_, int32_t searchRange, 
     torch::Tensor sortedPositions_, torch::Tensor sortedSupport_,
-    torch::Tensor hashTable_, int hashMapLength, 
+    torch::Tensor hashTable_, int32_t hashMapLength, 
     torch::Tensor numCells_, torch::Tensor cellTable_,
     torch::Tensor qMin_, float_t hCell, torch::Tensor maxDomain_, torch::Tensor minDomain_, torch::Tensor periodicity_,
     std::string mode, bool verbose = false){
@@ -286,7 +286,7 @@ std::pair<torch::Tensor, torch::Tensor> buildNeighborList_t(
     auto sortedSupport = getAccessor<float_t, 1>(sortedSupport_, "sortedSupport", useCuda, verbose, supportMode::gather == searchMode);
 
     // Check if the datastructure tensors are defined and contiguous and have the correct dimensions
-    auto hashTable = getAccessor<int, 2>(hashTable_, "hashTable", useCuda, verbose);
+    auto hashTable = getAccessor<int64_t, 2>(hashTable_, "hashTable", useCuda, verbose);
     auto numCells = getAccessor<int, 1>(numCells_, "numCells", useCuda, verbose);
     auto cellTable = getAccessor<int64_t, 2>(cellTable_, "cellTable", useCuda, verbose);
     auto qMin = getAccessor<float_t, 1>(qMin_, "qMin", useCuda, verbose);
@@ -296,13 +296,13 @@ std::pair<torch::Tensor, torch::Tensor> buildNeighborList_t(
 
 
     // Check if the neighbor counter tensor is defined and contiguous
-    auto neighborCounter = getAccessor<int, 1>(neighborCounter_, "neighborCounter", useCuda, verbose);
-    auto neighborOffsets = getAccessor<int, 1>(neighborOffsets_, "neighborOffsets", useCuda, verbose);
+    auto neighborCounter = getAccessor<int64_t, 1>(neighborCounter_, "neighborCounter", useCuda, verbose);
+    auto neighborOffsets = getAccessor<int64_t, 1>(neighborOffsets_, "neighborOffsets", useCuda, verbose);
 
     // Get the dimensions of the input tensors
-    int nQuery = queryPositions.size(0);
-    int dim = queryPositions.size(1);
-    int nSorted = sortedPositions.size(0);
+    int32_t nQuery = queryPositions.size(0);
+    int32_t dim = queryPositions.size(1);
+    int32_t nSorted = sortedPositions.size(0);
 
     
     auto periodicBoolHost = periodicity_.to(at::kCPU).to(at::kBool);
@@ -382,21 +382,22 @@ std::pair<torch::Tensor, torch::Tensor> buildNeighborList_t(
     }
 
     // Allocate output tensor for the neighbor counters
-    auto neighborList_i = torch::zeros({neighborListLength}, defaultOptions.dtype(torch::kInt32));
-    auto neighborList_j = torch::zeros({neighborListLength}, defaultOptions.dtype(torch::kInt32));
+    auto neighborList_i = torch::zeros({neighborListLength}, defaultOptions.dtype(torch::kInt64));
+    auto neighborList_j = torch::zeros({neighborListLength}, defaultOptions.dtype(torch::kInt64));
 
+    // return std::make_pair(neighborList_i, neighborList_j);
     // Create the accessors for the input tensors as packed accessors
     auto referencePositionAccessor = sortedPositions_.packed_accessor32<float_t, 2, traits>();
     auto referenceSupportAccessor = sortedSupport_.packed_accessor32<float_t, 1, traits>();
-    auto hashTableAccessor = hashTable_.packed_accessor32<int32_t, 2, traits>();
+    auto hashTableAccessor = hashTable_.packed_accessor32<int64_t, 2, traits>();
     auto cellTableAccessor = cellTable_.packed_accessor32<int64_t, 2, traits>();
     auto offsetAccessor = offsets.packed_accessor32<int32_t, 2, traits>();
     auto numCellsAccessor = numCells_.packed_accessor32<int32_t, 1, traits>();
-    auto neighborCounterAccessor = neighborCounter_.packed_accessor32<int32_t, 1, traits>();
-    auto neighborOffsetsAccessor = neighborOffsets_.packed_accessor32<int32_t, 1, traits>();
+    auto neighborCounterAccessor = neighborCounter_.packed_accessor32<int64_t, 1, traits>();
+    auto neighborOffsetsAccessor = neighborOffsets_.packed_accessor32<int64_t, 1, traits>();
 
-    auto neighborList_iAccessor = neighborList_i.packed_accessor32<int32_t, 1, traits>();
-    auto neighborList_jAccessor = neighborList_j.packed_accessor32<int32_t, 1, traits>();
+    auto neighborList_iAccessor = neighborList_i.packed_accessor32<int64_t, 1, traits>();
+    auto neighborList_jAccessor = neighborList_j.packed_accessor32<int64_t, 1, traits>();
 
     // Loop over all query particles and count the number of neighbors per particle
     // auto dim = queryPositions.size(1);
@@ -442,9 +443,9 @@ std::pair<torch::Tensor, torch::Tensor> buildNeighborList_t(
 
 // Define the python bindings for the C++ functions
 torch::Tensor countNeighbors(
-    torch::Tensor queryPositions_, torch::Tensor querySupport_, int searchRange, 
+    torch::Tensor queryPositions_, torch::Tensor querySupport_, int32_t searchRange, 
     torch::Tensor sortedPositions_, torch::Tensor sortedSupport_,
-    torch::Tensor hashTable_, int hashMapLength, 
+    torch::Tensor hashTable_, int32_t hashMapLength, 
     torch::Tensor numCells_, torch::Tensor cellTable_,
     torch::Tensor qMin_, double hCell, torch::Tensor maxDomain_, torch::Tensor minDomain_, torch::Tensor periodicity_,
     std::string mode, bool verbose){
@@ -462,10 +463,10 @@ torch::Tensor countNeighbors(
     }
 
 std::pair<torch::Tensor, torch::Tensor> buildNeighborList(
-    torch::Tensor neighborCounter_, torch::Tensor neighborOffsets_, int neighborListLength,
-    torch::Tensor queryPositions_, torch::Tensor querySupport_, int searchRange, 
+    torch::Tensor neighborCounter_, torch::Tensor neighborOffsets_, int32_t neighborListLength,
+    torch::Tensor queryPositions_, torch::Tensor querySupport_, int32_t searchRange, 
     torch::Tensor sortedPositions_, torch::Tensor sortedSupport_,
-    torch::Tensor hashTable_, int hashMapLength, 
+    torch::Tensor hashTable_, int32_t hashMapLength, 
     torch::Tensor numCells_, torch::Tensor cellTable_,
     torch::Tensor qMin_, double hCell, torch::Tensor maxDomain_, torch::Tensor minDomain_, torch::Tensor periodicity_,
     std::string mode, bool verbose){
