+++ src/torchCompactRadius/cppSrc/neighborhoodFixed.h
@@ -14,10 +14,10 @@
 
 template<std::size_t dim, typename scalar_t>
 hostDeviceInline auto countNeighborsForParticleFixed(int32_t i,
-    ptr_t<int32_t, 1> neighborCounters, 
-    cptr_t<scalar_t, 2> queryPositions, int searchRange, 
+    ptr_t<int64_t, 1> neighborCounters, 
+    cptr_t<scalar_t, 2> queryPositions, int32_t searchRange, 
     cptr_t<scalar_t, 2> sortedPositions, scalar_t support,
-    cptr_t<int32_t, 2> hashTable, int hashMapLength,
+    cptr_t<int64_t, 2> hashTable, int32_t hashMapLength,
     cptr_t<int64_t, 2> cellTable, cptr_t<int32_t,1> numCellsVec, 
     cptr_t<int32_t, 2> offsets,
     scalar_t hCell, cptr_t<scalar_t,1> minDomain, cptr_t<scalar_t,1> maxDomain, cptr_t<int32_t,1> periodicity){
@@ -26,7 +26,7 @@ hostDeviceInline auto countNeighborsForParticleFixed(int32_t i,
     // auto dim = xi.size(0);
     // auto queryCell = torch::zeros({dim}, defaultOptions.dtype(torch::kInt32));
     std::array<int32_t, dim> queryCell;
-    for(int d = 0; d < dim; d++)
+    for(int32_t d = 0; d < dim; d++)
         queryCell[d] = std::floor((xi[d] - minDomain[d]) / hCell);
     int32_t neighborCounter = 0;
     iterateOffsetCells(queryCell, offsets, 
@@ -47,10 +47,10 @@ hostDeviceInline auto countNeighborsForParticleFixed(int32_t i,
 
 template<std::size_t dim, typename scalar_t>
 hostDeviceInline auto buildNeighborhoodFixed(int32_t i,
-                       cptr_t<int32_t, 1> neighborOffsets, ptr_t<int32_t, 1> neighborList_i, ptr_t<int32_t, 1> neighborList_j,
-                       cptr_t<scalar_t, 2> queryPositions, int searchRange,
+                       cptr_t<int64_t, 1> neighborOffsets, ptr_t<int64_t, 1> neighborList_i, ptr_t<int64_t, 1> neighborList_j,
+                       cptr_t<scalar_t, 2> queryPositions, int32_t searchRange,
                        cptr_t<scalar_t, 2> sortedPositions, scalar_t support,
-                       cptr_t<int32_t, 2> hashTable, int hashMapLength,
+                       cptr_t<int64_t, 2> hashTable, int32_t hashMapLength,
                        cptr_t<int64_t, 2> cellTable, cptr_t<int32_t, 1> numCells,
                        cptr_t<int32_t, 2> offsets, scalar_t hCell, cptr_t<scalar_t, 1> minDomain, cptr_t<scalar_t, 1> maxDomain, cptr_t<int32_t, 1> periodicity) {
                         scalar_t h2 = support * support;
@@ -64,7 +64,7 @@ hostDeviceInline auto buildNeighborhoodFixed(int32_t i,
     // auto dim = xi.size(0);
     // auto queryCell = torch::zeros({dim}, defaultOptions.dtype(torch::kInt32));
     std::array<int32_t, dim> queryCell;
-    for(int d = 0; d < dim; d++)
+    for(int32_t d = 0; d < dim; d++)
         queryCell[d] = std::floor((xi[d] - minDomain[d]) / hCell);
 
     iterateOffsetCells(
@@ -86,16 +86,16 @@ hostDeviceInline auto buildNeighborhoodFixed(int32_t i,
 
 void countNeighborsForParticleCudaFixed(
     torch::Tensor neighborCounters, 
-    torch::Tensor queryPositions, int searchRange, 
+    torch::Tensor queryPositions, int32_t searchRange, 
     torch::Tensor sortedPositions, double support,
-    torch::Tensor hashTable, int hashMapLength,
+    torch::Tensor hashTable, int32_t hashMapLength,
     torch::Tensor cellTable, torch::Tensor numCellsVec, 
     torch::Tensor offsets,
     float hCell, torch::Tensor minDomain, torch::Tensor maxDomain, torch::Tensor periodicity) ;
 void buildNeighborhoodCudaFixed(
     torch::Tensor neighborOffsets, torch::Tensor neighborList_i, torch::Tensor neighborList_j,
-    torch::Tensor queryPositions, int searchRange,
+    torch::Tensor queryPositions, int32_t searchRange,
     torch::Tensor sortedPositions, double support,
-    torch::Tensor hashTable, int hashMapLength,
+    torch::Tensor hashTable, int32_t hashMapLength,
     torch::Tensor cellTable, torch::Tensor numCells,
     torch::Tensor offsets, double hCell, torch::Tensor minDomain, torch::Tensor maxDomain, torch::Tensor periodicity);
